fix: expand CookieJar type to mirror tough-cookie's full API
Moves the cookie types into their own file and rounds CookieJar out
with setCookie, getCookies*, getSetCookieStrings*, serialize*, so a
tough-cookie CookieJar is plug-and-play compatible with the interface.

Also updates docs/public-api/crawlee-types.api.md via `pnpm api:extract`.
