+++ packages/types/src/index.ts
@@ -2,6 +2,7 @@ export type * from './status-message.js';
 export type * from './storages.js';
 export type * from './utility-types.js';
 export type * from './browser.js';
+export type * from './cookies.js';
 export type * from './http-client.js';
 export type * from './session.js';
 export type * from './logger.js';
