+++ packages/types/src/session.ts
@@ -1,66 +1,4 @@
-/**
- * Cookie shape used by {@apilink CookieJar}. Structurally compatible with `tough-cookie`'s
- * `Cookie` class, so a `tough-cookie` cookie satisfies it without `@crawlee/types` depending
- * on `tough-cookie` itself.
- */
-export interface SessionCookie {
-    key: string;
-    value: string;
-    expires: Date | 'Infinity' | null;
-    maxAge: number | 'Infinity' | '-Infinity' | null;
-    domain: string | null;
-    path: string | null;
-    secure: boolean;
-    httpOnly: boolean;
-    extensions: string[] | null;
-    creation: Date | 'Infinity' | null;
-    creationIndex: number;
-    hostOnly: boolean | null;
-    pathIsDefault: boolean | null;
-    lastAccessed: Date | 'Infinity' | null;
-    sameSite: string | undefined;
-    toJSON(): Record<string, unknown>;
-    clone(): SessionCookie | undefined;
-    validate(): boolean;
-    setExpires(exp: string | Date): void;
-    setMaxAge(age: number): void;
-    cookieString(): string;
-    toString(): string;
-    TTL(now?: number): number;
-    expiryTime(now?: Date): number | undefined;
-    expiryDate(now?: Date): Date | undefined;
-    isPersistent(): boolean;
-    canonicalizedDomain(): string | undefined;
-    cdomain(): string | undefined;
-}
-
-/**
- * Minimal cookie jar contract, covering only what Crawlee itself calls. Structurally
- * compatible with `tough-cookie`'s `CookieJar`, which is what backs {@apilink Session} by
- * default, but `@crawlee/types` does not depend on `tough-cookie` for it.
- */
-export interface CookieJar {
-    setCookieSync(
-        cookie: string | SessionCookie,
-        url: string,
-        options?: { ignoreError?: boolean },
-    ): SessionCookie | undefined;
-    getCookiesSync(url: string): SessionCookie[];
-    getCookieStringSync(url: string): string;
-    clone(): Promise<CookieJar>;
-}
-
-/**
- * JSON representation of a {@apilink CookieJar}. Structurally compatible with `tough-cookie`'s
- * `SerializedCookieJar`.
- */
-export interface SerializedCookieJar {
-    version: string;
-    storeType: string | null;
-    rejectPublicSuffixes: boolean;
-    cookies: Record<string, unknown>[];
-    [key: string]: unknown;
-}
+import type { CookieJar, SerializedCookieJar } from './cookies.js';
 
 /**
  * The main purpose of the ProxyInfo object is to provide information
